+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.2.6] / 2023-03-28
+### Features
+- [] Tests with abstract class with [Test] methods.
+- [x] Tests alias with arguments show type parameters.
+- [x] Test with parameters works with IsAssignableFrom.
+
 ## [1.2.5] / 2023-03-27
 ### Fixed
 - Fix `Explicit` not working, add test.
@@ -164,6 +170,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Test with Console Output
 
 [vNext]: ../../compare/1.0.0...HEAD
+[1.2.6]: ../../compare/1.2.5...1.2.6
 [1.2.5]: ../../compare/1.2.4...1.2.5
 [1.2.4]: ../../compare/1.2.3...1.2.4
 [1.2.3]: ../../compare/1.2.2...1.2.3